third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-workers-bypass-for-network-redirect.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="service-workers-test.js"></script>
<script src="../resources-test.js"></script>
<script>

function loadIframe()
{
    var frame = document.createElement('iframe');
    frame.src = "http://localhost:8000/inspector/service-workers/resources/" +
                "bypass-for-network-redirect.php"
    document.body.appendChild(frame);
}

function test()
{
    UI.inspectorView.showPanel("sources")
        .then(function(){
            Common.settings.settingForTest("bypassServiceWorker").set(true);
            var callback;
            var promise = new Promise((fulfill) => callback = fulfill);
            ConsoleTestRunner.addConsoleSniffer(message => {
                if (message.messageText == "getRegistration finished") {
                  callback();
                }
              }, true);
            TestRunner.evaluateInPage("loadIframe()");
            return promise;
        })
        .then(function() {
            TestRunner.addResult("Success");
            TestRunner.completeTest();
        })
        .catch(function(exception) {
            TestRunner.addResult("Error");
            TestRunner.addResult(exception);
            TestRunner.completeTest();
        });
}

</script>
</head>
<body onload="runTest()">
<p>Tests "Bypass for network" checkbox with redirection doesn't cause crash.<p>
</body>
</html>
